i18n

Macro tr

tr!() { /* proc-macro */ }

Queries for a localized message, returning an i18n::Message instance.

This macro provides a convenient way to query for a message using a language identifier, message ID, and optional arguments. It handles error cases by providing a default fallback message.

Syntax

tr!(lang: Expr, id: LitStr [, locales = VAR_NAME] [, key = value]* [, .attribute_name(key = value)* ])

lang: A Rust expression that evaluates to a &LanguageIdentifier (e.g., langid!("en-US") or a variable). This is the language to query for.

id: A string literal representing the ID of the Fluent message.

locales (optional): An identifier for the i18n::Locales static variable to use. Defaults to LOCALES.

key = value: Optional key-value pairs for arguments to the main message. key must be an string literal, and value can be any Rust expression.

attr(attr_id, key = value): Optional arguments for a specific attribute of the message. attr_id is a string literal representing the attribute ID (e.g., “aria-label” or “attr-arg”).

Returns

An i18n::Message instance.

Examples

// Load the localization data.
i18n::load!("../tests/i18n", fallback_lang = "en-US", name = TR_LOCALES);

let lang_en = i18n::langid!("en-US");

// Basic usage:
let msg1 = i18n::tr!(lang_en, "foo-a", locales = TR_LOCALES);
assert_eq!(msg1.value, "English A".to_string());

// With main message arguments:
let msg2 = i18n::tr!(lang_en, "welcome-back", "username" = "Alice", locales = TR_LOCALES);
assert_eq!(msg2.value, "Welcome back, \u{2068}Alice\u{2069}!".to_string());

// With attribute arguments:
let msg3 = i18n::tr!(lang_en, "login-btn", attr("attr-arg", "text" = "some text"), locales = TR_LOCALES);
assert_eq!(msg3.attrs.get("attr-arg"), Some(&"This is an attribute argument with arbitrary text: \u{2068}some text\u{2069}".to_string()));

// With a custom locales variable:
i18n::load!("../tests/i18n", fallback_lang = "en-US", name = MY_APP_LOCALES);
let msg4 = i18n::tr!(lang_en, "foo-b", locales = MY_APP_LOCALES);
assert_eq!(msg4.value, "English B".to_string());